+++ CHANGES.rst
@@ -79,8 +79,9 @@ Unreleased
         allows the user to search for future output of the generator when
         using less and then aborting the program using ctrl-c.
 
--   Show deprecation warning message running ``--help`` Option
-    for ``@click.option``. :issue:`2263`
+-   Parameters can now be deprecated using ``deprecated=True``. These cannot be
+    required nor prompted and are highlighted in usage output. Using them
+    will result in a warning being printed. :issue:`2263` :pr:`2271`
 
 
 Version 8.1.8